adding tests for unfoldEvalF
